feat: Add Header, Footer, HeroSection, and HomeLayout components
- Implemented Header component with user authentication and theme settings.
- Created Footer component with company info, product links, resources, and newsletter signup.
- Developed HeroSection component featuring an interactive textbox for user input.
- Established HomeLayout component to structure the main layout with Header, HeroSection, and Footer.
- Added index file for easier exports of home components.
