+++ components/home/index.ts
@@ -1,0 +1,4 @@
+export { default as Header } from "./Header";
+export { default as HeroSection } from "./HeroSection";
+export { default as Footer } from "./Footer";
+export { default as HomeLayout } from "./HomeLayout";
